+++ Package@swift-6.0.swift
@@ -78,6 +78,13 @@ let package = Package(
                 "SwiftXMLCoderTestSupport",
                 "SwiftXMLCoderMacros"
             ]
+        ),
+        .testTarget(
+            name: "SwiftXMLCoderMacroTests",
+            dependencies: [
+                "SwiftXMLCoderMacroImplementation",
+                .product(name: "SwiftSyntaxMacrosTestSupport", package: "swift-syntax")
+            ]
         )
     ],
     swiftLanguageModes: [